added feedback.txt file and thank you html page. Also added data collection route to App.py which allows for submision of data to feedback.txt
fixed error in data required where i forgot to put () at the end of it, also added data collection route that will take data added to form and put it in feedback.txt file which was created and added to the project. also made thank you html page and route for it which it will take you to when form is successfuly submitted.